+++ StudCity.Frontend/src/App.js
@@ -7,7 +7,7 @@ import {
 import Main from "./pages/Main";
 import Authenticate from "./pages/authenticate";
 import Registration from "./pages/registration";
-import VerifyEmail from "./pages/verify-email/VerifyEmail";
+import VerifyEmail from "./pages/verify-email";
 
 const router = createBrowserRouter([
 	{
